Added fire missile function
