+++ src/key_pressed.h
@@ -1,0 +1,2 @@
+char key_pressed();
+void fireMissile(char missile, int y, int x);
