+++ src/main.c
@@ -5,26 +5,8 @@
 #include <termios.h>
 #include <unistd.h>
 #include <fcntl.h>
-
-extern char *GetShip(char *fileName, int *fileSizePTR);
-extern void diplayShip(int fileSizeShip, char *shipFile, int y, int x);
-void eraseShip (int fileSizeShip, char *shipFile, int y, int x);
-
-char key_pressed()
-{
-struct termios oldterm, newterm;
-int oldfd; char c, result = 0;
-tcgetattr (STDIN_FILENO, &oldterm);
-newterm = oldterm; newterm.c_lflag &= ~(ICANON | ECHO);
-tcsetattr (STDIN_FILENO, TCSANOW, &newterm);
-oldfd = fcntl (STDIN_FILENO, F_GETFL, 0);
-fcntl (STDIN_FILENO, F_SETFL, oldfd | O_NONBLOCK) ;
-c = getchar();
-tcsetattr (STDIN_FILENO, TCSANOW, &oldterm);
-fcntl (STDIN_FILENO, F_SETFL, oldfd);
-if (c != EOF) {ungetc(c, stdin) ; result = getchar();}
-return result;
-}
+#include "struct_def.h"
+#include "key_pressed.h"
 
 int main(void){
   system("clear");
@@ -36,45 +18,70 @@ int main(void){
   printf("\033[%d;%dH", y, x);
   diplayShip(fileSizeShip1, shipFile1, y, x);
   char a=0;
+  char missile=73;
+  int j=0;
+  int fire=0;
+  int t, u;
   printf("\e[?25l"); //hide the terminal cursor
 
   while (1)
   {
-  a=key_pressed();
-    if (a != 0)
+  if (fire==1)
+  {
+    fireMissile(missile, t+j, u);
+    j++;
+    if (j>50)
     {
-      if (a == 122)//Up - Z key
-    {
-      eraseShip(fileSizeShip1, shipFile1, y, x);
-      y=y-4;
-      printf("\033[%d;%dH", y, x);
-      diplayShip(fileSizeShip1, shipFile1, y, x);
-      printf("\033[%d;%dH", y, x);
+      fire=0;
+      j=0;
     }
-    if (a == 115)//Down - S key
+  }
+    a=key_pressed();
+    if (a != 0) //Ship movement
     {
-      eraseShip(fileSizeShip1, shipFile1, y, x);
-      y=y+4;
-      printf("\033[%d;%dH", y, x);
-      diplayShip(fileSizeShip1, shipFile1, y, x);
-      printf("\033[%d;%dH", y, x);
-    }
-    if (a == 113)//Left - Q key
-    {
-      eraseShip(fileSizeShip1, shipFile1, y, x);
-      x=x-4;
-      printf("\033[%d;%dH", y, x);
-      diplayShip(fileSizeShip1, shipFile1, y, x);
-      printf("\033[%d;%dH", y, x);
-    }
-    if (a == 100)//Right - D key
-    {
-      eraseShip(fileSizeShip1, shipFile1, y, x);
-      x=x+4;
-      printf("\033[%d;%dH", y, x);
-      diplayShip(fileSizeShip1, shipFile1, y, x);
-      printf("\033[%d;%dH", y, x);
-    }
+        if (a == 122)//Up - Z key
+      {
+        eraseShip(fileSizeShip1, shipFile1, y, x);
+        y=y-4;
+        printf("\033[%d;%dH", y, x);
+        diplayShip(fileSizeShip1, shipFile1, y, x);
+        printf("\033[%d;%dH", y, x);
+      }
+      if (a == 115)//Down - S key
+      {
+        eraseShip(fileSizeShip1, shipFile1, y, x);
+        y=y+4;
+        printf("\033[%d;%dH", y, x);
+        diplayShip(fileSizeShip1, shipFile1, y, x);
+        printf("\033[%d;%dH", y, x);
+      }
+      if (a == 113)//Left - Q key
+      {
+        eraseShip(fileSizeShip1, shipFile1, y, x);
+        x=x-4;
+        printf("\033[%d;%dH", y, x);
+        diplayShip(fileSizeShip1, shipFile1, y, x);
+        printf("\033[%d;%dH", y, x);
+      }
+      if (a == 100)//Right - D key
+      {
+        eraseShip(fileSizeShip1, shipFile1, y, x);
+        x=x+4;
+        printf("\033[%d;%dH", y, x);
+        diplayShip(fileSizeShip1, shipFile1, y, x);
+        printf("\033[%d;%dH", y, x);
+      }
+      if (a == 32)//Fire missile - Space key
+      {
+        if (fire==0)
+        {
+        t=y;
+        u=x;
+        t=t+5;
+        u=u+4;
+        fire=1;
+        }       
+      }
     }
   }
   //diplayShip(fileSizeShip2, shipFile2);
