+++ src/struct_def.h
@@ -1,0 +1,3 @@
+char *GetShip(char *fileName, int *fileSizePTR);
+void diplayShip(int fileSizeShip, char *shipFile, int y, int x);
+void eraseShip (int fileSizeShip, char *shipFile, int y, int x);
